+++ src/common/functools.ts
